# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Firat Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/dateofintern/all": {
            "get": {
                "tags": [
                    "Staj Tarihi"
                ],
                "description": "Staj Tarihlerini Getir",
                "operationId": "9a592c284fc324c3acb7d288f94030f9",
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/v1/dateofintern/get": {
            "get": {
                "tags": [
                    "Staj Tarihi"
                ],
                "description": "Staj Tarihlerini id'ye Göre Getir",
                "operationId": "ea0d911e9d904b6474e605b2ca8de21d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Parametre açıklaması",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/dateofintern/create": {
            "post": {
                "tags": [
                    "Staj Tarihi"
                ],
                "description": "Staj Tarihi Oluştur",
                "operationId": "99c42960c7cff73c7bac9a4099ab810b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start_date",
                                    "expire_date"
                                ],
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "date_format": "m/d/Y"
                                    },
                                    "expire_date": {
                                        "type": "string",
                                        "format": "date",
                                        "date_format": "m/d/Y"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/v1/dateofintern/update": {
            "post": {
                "tags": [
                    "Staj Tarihi"
                ],
                "description": "Staj Tarihi Güncelle",
                "operationId": "7d4a5b8094ba87772eb93eccc9ce0d7e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "date_format": "m/d/Y"
                                    },
                                    "expire_date": {
                                        "type": "string",
                                        "format": "date",
                                        "date_format": "m/d/Y"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/v1/dateofintern/delete": {
            "post": {
                "tags": [
                    "Staj Tarihi"
                ],
                "description": "Staj Tarihi Sil",
                "operationId": "f369cbb276d0b83a0d0baf7f29aeda07",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
...
    }
}